+++ LLIO2025/package.json
@@ -8,16 +8,16 @@
 		"build": "vite build",
 		"preview": "vite preview",
 		"prepare": "svelte-kit sync || echo ''",
-    "format": "prettier --write --plugin prettier-plugin-svelte .",
-    "format:check": "prettier --check --plugin prettier-plugin-svelte ."
+		"format": "prettier --write --plugin prettier-plugin-svelte .",
+		"format:check": "prettier --check --plugin prettier-plugin-svelte ."
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^4.0.0",
 		"@sveltejs/adapter-node": "^5.2.12",
 		"@sveltejs/kit": "^2.16.0",
 		"@sveltejs/vite-plugin-svelte": "^5.0.0",
-    "prettier": "^3.5.0",
-    "prettier-plugin-svelte": "^3.3.3",
+		"prettier": "^3.5.0",
+		"prettier-plugin-svelte": "^3.3.3",
 		"svelte": "^5.0.0",
 		"vite": "^6.0.0"
 	},
@@ -25,6 +25,8 @@
 		"@fullcalendar/core": "^6.1.15",
 		"@fullcalendar/daygrid": "^6.1.15",
 		"@fullcalendar/interaction": "^6.1.15",
-		"@fullcalendar/timegrid": "^6.1.15"
+		"@fullcalendar/timegrid": "^6.1.15",
+		"@tailwindcss/vite": "^4.0.9",
+		"tailwindcss": "^4.0.9"
 	}
 }
