--- LLIO2025/src/routes/login/+page.svelte
@@ -1,18 +1,0 @@
-<div class="wrapper">
-  <form>
-    <h1>Login</h1>
-    <div class="input-box">
-      <input type="text" placeholder="Email" required />
-    </div>
-    <div class="input-box">
-      <input type="password" placeholder="Mot de passe" required />
-    </div>
-    <div class="remember-forgot">
-      <label>
-        <input type="checkbox" name="remember" value="remember" />
-        Se souvenir de moi
-      </label>
-    </div>
-    <button type="submit">Se connecter</button>
-  </form>
-</div>
