+++ LLIO2025/src/style/app.css
@@ -1,0 +1,1 @@
+@import "tailwindcss";
