+++ LLIO2025/vite.config.js
@@ -1,8 +1,12 @@
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
+import tailwindcss from '@tailwindcss/vite'
 
 export default defineConfig({
-  plugins: [sveltekit()],
+  plugins: [
+    tailwindcss(),
+    sveltekit()
+  ],
   server: {
     allowedHosts: ['dev.edwrdl.ca'],
   },
